refactor: postDto 변수명 수정, summary 서브 패키지 생성
