+++ blog/src/main/java/com/practice/blog/account/dto/response/AccountResponseDto.java
@@ -1,21 +1,20 @@
-package com.practice.blog.account.dto;
+package com.practice.blog.account.dto.response;
 
 import com.practice.blog.account.entity.Account;
 import lombok.Builder;
 import lombok.Getter;
 
-// Account 생성 후 응답 DTO
+// Account 조회 후 응답 DTO
 
 @Builder @Getter
-public class CreateAccountResponseDto {
-    private Long id;
+public class AccountResponseDto {
+
     private String nickname;
     private String email;
     private String bio;
 
-    public static CreateAccountResponseDto from(Account account) {
-        return CreateAccountResponseDto.builder()
-                .id(account.getAccountId())
+    public static AccountResponseDto from(Account account) {
+        return AccountResponseDto.builder()
                 .nickname(account.getNickname())
                 .email(account.getEmail())
                 .bio(account.getBio())
