+++ blog/src/main/java/com/practice/blog/account/dto/response/CreateAccountResponseDto.java
@@ -1,20 +1,21 @@
-package com.practice.blog.account.dto;
+package com.practice.blog.account.dto.response;
 
 import com.practice.blog.account.entity.Account;
 import lombok.Builder;
 import lombok.Getter;
 
-// Account 조회 후 응답 DTO
+// Account 생성 후 응답 DTO
 
 @Builder @Getter
-public class AccountResponseDto {
-
+public class CreateAccountResponseDto {
+    private Long id;
     private String nickname;
     private String email;
     private String bio;
 
-    public static AccountResponseDto from(Account account) {
-        return AccountResponseDto.builder()
+    public static CreateAccountResponseDto from(Account account) {
+        return CreateAccountResponseDto.builder()
+                .id(account.getAccountId())
                 .nickname(account.getNickname())
                 .email(account.getEmail())
                 .bio(account.getBio())
