+++ blog/src/main/java/com/practice/blog/account/dto/response/ErrorResponseDto.java
@@ -1,9 +1,7 @@
-package com.practice.blog.account.dto;
+package com.practice.blog.account.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
