+++ blog/src/main/java/com/practice/blog/post/controller/PostController.java
@@ -1,10 +1,9 @@
 package com.practice.blog.post.controller;
 
-import com.practice.blog.post.domain.Post;
 import com.practice.blog.post.dto.request.PostCreateRequest;
 import com.practice.blog.post.dto.request.PostUpdateRequest;
 import com.practice.blog.post.dto.response.PostResponse;
-import com.practice.blog.post.dto.response.PostsResponses;
+import com.practice.blog.post.dto.response.PostListResponse;
 import com.practice.blog.post.service.PostService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +27,7 @@ public class PostController {
 
     // 게시물 목록 조회
     @GetMapping
-    public ResponseEntity<PostsResponses> getAllPosts(){
+    public ResponseEntity<PostListResponse> getAllPosts(){
         return ResponseEntity.ok(postService.getAllPosts());
     }
