+++ blog/src/main/java/com/practice/blog/post/dto/request/PostUpdateRequest.java
@@ -2,5 +2,9 @@ package com.practice.blog.post.dto.request;
 
 import jakarta.validation.constraints.Size;
 
-public record PostUpdateRequest(@Size(min=5, max=500, message = "내용은 5자이상 500자이하로 입력해야합니다.") String content) {
+public record PostUpdateRequest(
+        @Size(min = 5, max = 500, message = "내용은 5자 이상 500자 이하로 입력해야 합니다.")
+        String content
+) {
 }
+
