+++ blog/src/main/java/com/practice/blog/post/dto/response/PostListResponse.java
@@ -1,0 +1,8 @@
+package com.practice.blog.post.dto.response;
+
+import com.practice.blog.post.dto.summary.PostSummary;
+
+import java.util.List;
+
+public record PostListResponse(List<PostSummary> posts, Long totalPosts) {}
+
