+++ blog/src/main/java/com/practice/blog/post/dto/response/PostResponse.java
@@ -4,6 +4,7 @@ import com.practice.blog.post.domain.Post;
 
 import java.time.LocalDateTime;
 
+// 단건 조회
 public record PostResponse(Long postId, Long accountId, String nickName, String title, String content,
                            LocalDateTime createdAt, LocalDateTime modifiedAt, Long viewCount) {
     public static PostResponse from(Post post) {
