--- blog/src/main/java/com/practice/blog/post/dto/response/PostsResponses.java
@@ -1,6 +1,0 @@
-package com.practice.blog.post.dto.response;
-
-import java.util.List;
-
-public record PostsResponses(List<PostsResponse> posts, Long totalPosts) {
-}
